requirements.txt
whisper-timestamped
transformers
django
google-api-python-client
google-auth-oauthlib
google-auth-httplib2